requirements.txt
annotated-types==0.5.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
pydantic==2.0.2
pydantic_core==2.1.2
pytest==7.4.0
pytest-env==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vcrpy==5.0.0
wrapt==1.15.0
yarl==1.9.2
